# ctapipe.image.cleaning.tailcuts_clean speed up factor ~10