--- a/ctapipe/image/cleaning.py
+++ b/ctapipe/image/cleaning.py
@@ -30,11 +30,11 @@ def tailcuts_clean(geom, image, picture_thresh=7, boundary_thresh=5,
     picture_thresh: float or array
         threshold above which all pixels are retained
     boundary_thresh: float or array
-        threshold above which pixels are retained if they have a neighbor 
+        threshold above which pixels are retained if they have a neighbor
         already above the picture_thresh
     keep_isolated_pixels: bool
-        If True, pixels above the picture threshold will be included always, 
-        if not they are only included if a neighbor is in the picture or 
+        If True, pixels above the picture threshold will be included always,
+        if not they are only included if a neighbor is in the picture or
         boundary
     min_number_picture_neighbors: int
         A picture pixel survives cleaning only if it has at least this number
@@ -49,40 +49,45 @@ def tailcuts_clean(geom, image, picture_thresh=7, boundary_thresh=5,
     `image[~mask] = 0`
 
     """
-    pixels_above_picture = image >= picture_thresh
+    # core_pixels: I call pixels, which are above the so called "picture_thresh"
+    # edge_pixels: I call pixels, which are above the so called "boundary_thresh"
+
+
+    core_pixels = image >= picture_thresh
 
     if keep_isolated_pixels or min_number_picture_neighbors == 0:
-        pixels_in_picture = pixels_above_picture
+        pixels_in_picture = core_pixels
     else:
         # Require at least min_number_picture_neighbors. Otherwise, the pixel
         #  is not selected
-        number_of_neighbors_above_picture = np.sum(pixels_above_picture &
-                                                   geom.neighbor_matrix, axis=1)
-        pixels_in_picture = pixels_above_picture & (
-            number_of_neighbors_above_picture >= min_number_picture_neighbors
+        number_of_core_neighbors = np.sum(
+            core_pixels & geom.neighbor_matrix,
+            axis=1
         )
+        enough_core_neighbors = number_of_core_neighbors >= min_number_picture_neighbors
+        pixels_in_picture = core_pixels & enough_core_neighbors
 
     # by broadcasting together pixels_in_picture (1d) with the neighbor
     # matrix (2d), we find all pixels that are above the boundary threshold
     # AND have any neighbor that is in the picture
-    pixels_above_boundary = image >= boundary_thresh
+    edge_pixels = image >= boundary_thresh
     pixels_with_picture_neighbors = (pixels_in_picture &
                                      geom.neighbor_matrix).any(axis=1)
 
     if keep_isolated_pixels:
-        return (pixels_above_boundary
+        return (edge_pixels
                 & pixels_with_picture_neighbors) | pixels_in_picture
     else:
-        pixels_with_boundary_neighbors = (pixels_above_boundary &
+        pixels_with_boundary_neighbors = (edge_pixels &
                                           geom.neighbor_matrix).any(axis=1)
 
-        return ((pixels_above_boundary & pixels_with_picture_neighbors) |
+        return ((edge_pixels & pixels_with_picture_neighbors) |
                 (pixels_in_picture & pixels_with_boundary_neighbors))
 
 
 def dilate(geom, mask):
     """
-    Add one row of neighbors to the True values of a pixel mask and return 
+    Add one row of neighbors to the True values of a pixel mask and return
     the new mask.
     This can be used to include extra rows of pixels in a mask that was
     pre-computed, e.g. via `tailcuts_clean`.
@@ -91,7 +96,7 @@ def dilate(geom, mask):
     ----------
     geom: `~ctapipe.instrument.CameraGeometry`
         Camera geometry information
-    mask: ndarray 
+    mask: ndarray
         input mask (array of booleans) to be dilated
     """
     return mask | (mask & geom.neighbor_matrix).any(axis=1)