We have a kind of default analysis in digicampipe, which uses `ctapipe.image.cleaning.tailcuts_clean`. I realized this part takes ~13% of the total runtime of the analysis, while it is not operating on the waveforms (n_pixel, n_samples) but on the image, which is a factor of 50x smaller than the waveforms. 

Typically runtime scales somehow with N ... so I was surprised this was so slow. I started digging and saw that tailcut_clean performs basically dot-products with the neighbor matrix. They are not really written as `numpy.dot`-products, but they essentially are dot-products.

Since the neighbor-matrix is sparse by design these dots products can be sped by using that fact.
sparsity of neighbor matrices:
```
Name - fraction of non-zero entries:
---------------------------------------------
ASTRICam 0.16%
HESS-II 0.28%
Whipple490 0.89%
Whipple109 4.85%
CHEC 0.19%
FlashCam 0.33%
Whipple151 3.58%
Whipple331 1.70%
DigiCam 0.45%
HESS-I 0.60%
LSTCam 0.31%
VERITAS 1.14%
NectarCam 0.31%
SCTCam 0.03%
FACT 0.40%
```

In this PR I have just made a fast version of `tailcut_clean` alongside the normal version and implemented a few tests showing that both versions give still the same results and that the version, which claims to be faster is indeed faster.

I found this call to show it quite nicely:
```
 pytest ctapipe/image/tests/test_cleaning.py --durations=10
```

The result on my machine is:
```
dneise@lair:~/cta/ctapipe$ pytest ctapipe/image/tests/test_cleaning.py --durations=10
================================================= test session starts =================================================
platform linux -- Python 3.6.3, pytest-3.3.2, py-1.5.2, pluggy-0.6.0 -- /home/dneise/anaconda3/bin/python
cachedir: .cache
benchmark: 3.1.1 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=1.0 calibration_precision=10 warmup=False warmup_iterations=100000)
rootdir: /home/dneise/cta/ctapipe, inifile: setup.cfg
plugins: profiling-1.2.11, benchmark-3.1.1
collected 9 items                                                                                                     

ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_simple PASSED                                         [ 11%]
ctapipe/image/tests/test_cleaning.py::test_dilate PASSED                                                        [ 22%]
ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean PASSED                                                [ 33%]
ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_min_neighbors_1 PASSED                                [ 44%]
ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_min_neighbors_2 PASSED                                [ 55%]
ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_with_isolated_pixels PASSED                           [ 66%]
ctapipe/image/tests/test_cleaning.py::test_tailcuts_fast_equal_to_normal PASSED                                 [ 77%]
ctapipe/image/tests/test_cleaning.py::test_1000_tailcuts PASSED                                                 [ 88%]
ctapipe/image/tests/test_cleaning.py::test_1000_tailcuts_fast PASSED                                            [100%]

============================================== slowest 10 test durations ==============================================
2.09s call     ctapipe/image/tests/test_cleaning.py::test_1000_tailcuts
0.14s call     ctapipe/image/tests/test_cleaning.py::test_1000_tailcuts_fast
0.05s call     ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_simple
0.04s call     ctapipe/image/tests/test_cleaning.py::test_tailcuts_fast_equal_to_normal
0.03s call     ctapipe/image/tests/test_cleaning.py::test_dilate
0.03s call     ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean
0.00s call     ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_min_neighbors_1
0.00s call     ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_min_neighbors_2
0.00s call     ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_with_isolated_pixels
0.00s setup    ctapipe/image/tests/test_cleaning.py::test_tailcuts_clean_simple
============================================== 9 passed in 2.63 seconds ===============================================
```

## Things, I think are still to do:

-----

### Undo variable renaming:

Please note, I also renamed a couple of variable before I started to work, since I could not wrap my head around the meaning ... in hindsight this was stupid and I should rework this PR without the renaming. 

### Remove the duplication ...
... of `tailcuts_clean_fast` and `tailcuts_clean` there should be only one version in the end. 

### Remove the tests...
... showing that the fast version is indeed faster, since in the end there will probably be only the faster version.


----

Generally I am interested to hear your opinion about performance improving PRs like this one? 
Are they already wanted? Or is this premature?

And how do you like this style, where for a certain time both versions co-exist to test and compare them against each other and then after some time we decide to drop one version... is this the right way to go?

Looking forward to your reviews.