@dneise 2018-03-07
In another PR #666 I opened a cleaner version of this PR, so I'll close this one. Thanks @kosack for the review so far.